pull logic from theme change into a custom react hook
